Moved erc20deposit to new file
